Test serial
There are several bugs I found in `interrupts::tests`. Because of a
misconception, any time a `assert_throws` macro would be run more than
once, it would pass with the result of the previous run. This has been
fixed.

However, after fixing this bug, `test_no_interrupts` began to
triple-fault. It has been commented out and must be resolved.

resolves #45